read multichannel xdf.
TODO: markers?

+++ Transformations/IIRFilter/IIRFilter.m
@@ -50,7 +50,7 @@ if strcmp(options.useTable, 'Off')
             for chan = 1:nchan
                 TransTools.progressbar(chan/nchan);
                 drawnow;
-                output.data(chan,:,seg)=Tools.filtfilt(b,a,double(output.data(chan,:,seg)));
+                output.data(chan,:,seg)=Tools.filtfilt(b,a,output.data(chan,:,seg));
             end
         end
     end
@@ -71,13 +71,13 @@ else % Individual channel Setting:
                 [b,a] = TransTools.CreateFilter('low', input.srate, options.TableData.LCFreqValues(chan), str2double(char(options.TableData.LCSlopeValues(chan))));
                 TransTools.progressbar(chan/nchan);
                 drawnow;
-                output.data(chan,:,seg)=Tools.filtfilt(b,a,double(output.data(chan,:,seg)));
+                output.data(chan,:,seg)=Tools.filtfilt(b,a,output.data(chan,:,seg));
             end
             if options.TableData.HCSelected(chan)
                 [b,a] = TransTools.CreateFilter('high', input.srate, options.TableData.HCFreqValues(chan), str2double(char(options.TableData.HCSlopeValues(chan))));
                 TransTools.progressbar(chan/nchan);
                 drawnow;
-                output.data(chan,:,seg)=Tools.filtfilt(b,a,double(output.data(chan,:,seg)));
+                output.data(chan,:,seg)=Tools.filtfilt(b,a,output.data(chan,:,seg));
             end
         end
     end
